removed redundant headers, moved screens_init() from main.cpp to screens.cpp
