+++ client/src/main.cpp
@@ -1,44 +1,12 @@
-#include <SFML/Graphics.hpp>
 #include <TGUI/Backend/SFML-Graphics.hpp>
 #include <chrono>
 #include "../include/client.h"
-#include "../include/end_game_screen.h"
 #include "../include/game_object_size_constants.h"
-#include "../include/screen_defines.h"
 #include "../include/screens.h"
 #include "../include/sprite_printer.h"
 #include "../include/ui_functions.h"
 
 namespace war_of_ages {
-
-void screens_init(tgui::Gui &gui) {
-    // Widgets of every screen are defined here
-
-    // Ilya's screens
-
-    tournament_screen_init(gui);
-    tournament_creation_screen_init(gui);
-    tournament_join_screen_init(gui);
-
-    // end Ilya's screens
-
-    // Vakhtang's screens
-
-    start_screen_init(gui);
-    multiplayer_screen_init(gui);
-    settings_screen_init(gui);
-
-    // end Vakhtang's screens
-
-    // Timur's screens
-
-    opponent_waiting_screen_init(gui);
-    game_screen_init(gui);
-    end_game_screen_init(gui);
-    pause_screen_init(gui);
-
-    // end Timur's screens
-}
 
 war_of_ages::client_state current_state;
